--- src/mobgap/secondary_outcomes/__init__.py
@@ -1,5 +1,0 @@
-"""Methods to calculate secondary DMOs."""
-
-__all__ = ["SDMO"]
-
-from mobgap.secondary_outcomes._sdmo import SDMO
