+++ src/mobgap/signal_based/__init__.py
@@ -1,0 +1,5 @@
+"""Methods to calculate signal-based DMOs."""
+
+__all__ = ["SDMO"]
+
+from mobgap.signal_based._sdmo import SDMO
